% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cusip.R
\name{cusip_check}
\alias{cusip_check}
\title{Check validity of CUSIP}
\usage{
cusip_check(s)
}
\arguments{
\item{s}{a character vector for whose elements validity of CUSIP
is checked.}
}
\value{
A logical vector.
}
\description{
Given a character vector, check the validity of CUSIP
(Committee on Uniform Security Identification Procedures)
for each of its elements.
}
\examples{
cusip_check("052800109")
cusip_check("87162M409")
cusip_check("500750104")
cusip_check(c("052800109", "87162M409"))
cusip_check(c("052800109", "87162M407"))

}
