% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figi.R
\name{figi_check}
\alias{figi_check}
\title{Check validity of FIGI}
\usage{
figi_check(s)
}
\arguments{
\item{s}{a character vector for whose elements validity of FIGI
(Financial Instrument Global Identifier) is checked.}
}
\value{
A logical vector.
}
\description{
Given a character vector, check the validity of FIGI (Financial
Instrument Global Identifier) for each of its elements.
}
\examples{
figi_check("BBG000BLNQ16")
figi_check("NRG92C84SB39")
figi_check(c("BBG000BLNQ16", "NRG92C84SB39"))

}
