% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figi.R
\name{figi_compute_checksum}
\alias{figi_compute_checksum}
\title{Compute the FIGI check digit}
\usage{
figi_compute_checksum(v)
}
\arguments{
\item{v}{a character vector for whose elements the FIGI
(Financial Instrument Global Identifier) checksum digit is computed.}
}
\value{
A character vector of single character strings.
}
\description{
Given a character vector, compute for each string the
FIGI (Financial Instrument Global Identifier) checksum digit.
}
\examples{
figi_compute_checksum("BBG000BLNQ1")
figi_compute_checksum("NRG92C84SB3")
figi_compute_checksum(c("BBG000BLNQ1", "NRG92C84SB3"))

}
