% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isin.R
\name{isin_check}
\alias{isin_check}
\title{Check validity of ISIN}
\usage{
isin_check(s)
}
\arguments{
\item{s}{a character vector for whose elements validity of ISIN
(International Securities Identification Number) is checked.}
}
\value{
A logical vector.
}
\description{
Given a character vector, check the validity of ISIN
(International Securities Identification Number)
for each of its elements.
}
\examples{
isin_check("BBG000BLNQ16")
isin_check("NRG92C84SB39")
isin_check(c("BBG000BLNQ16", "NRG92C84SB39"))

}
