% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sedol.R
\name{sedol_check}
\alias{sedol_check}
\title{Check validity of SEDOL}
\usage{
sedol_check(s)
}
\arguments{
\item{s}{a character vector for whose elements validity of SEDOL
is checked.}
}
\value{
A logical vector.
}
\description{
Given a character vector, check the validity of SEDOL
(Stock Exchange Daily Official List)
for each of its elements.
}
\examples{
sedol_check('B014635')
sedol_check('0263494')
sedol_check(c('B014635', '0263494'))
sedol_check(c('B014635', '0263495'))

}
